model-service/requirements.txt
Flask
Flask-SQLAlchemy
bcrypt
PyJWT
flask-cors
python-dotenv
psycopg2-binary
pymongo
requests